+++ src/scripts/axios.ts
@@ -1,5 +1,6 @@
+import { config } from "@/config";
 import axios from "axios";
 
 export default axios.create({
-  baseURL: "http://localhost:8000",
+  baseURL: config.base,
 });