+++ src/scripts/store.ts
@@ -1,5 +1,5 @@
 import { defineStore } from "pinia";
-import { ref } from "vue";
+import { computed, ref } from "vue";
 
 const prefersDarkMode = () => {
   return (
@@ -28,7 +28,74 @@ export const useThemeStore = defineStore(
         document.documentElement.classList.remove("dark");
       }
     };
-    return { dark, toggle, init };
+    const icon = computed(() => (dark.value ? "pi pi-moon" : "pi pi-sun"));
+    const logo = computed(() => (dark.value ? "/acm-light.png" : "/acm.png"));
+    return { dark, toggle, init, icon, logo };
+  },
+  {
+    persist: true,
+  }
+);
+
+enum Role {
+  SuperAdmin = "super_admin",
+  Admin = "admin",
+  User = "user",
+  Inactive = "inactive",
+}
+
+interface Account {
+  id?: string;
+  token?: string;
+
+  username?: string;
+  email?: string;
+  avatar?: string;
+  signature?: string;
+  links?: string[];
+
+  nickname?: string;
+  sex?: boolean;
+  birthday?: Date;
+
+  name?: string;
+  student_id?: string;
+  school?: string;
+  college?: string;
+  major?: string;
+
+  rating?: number;
+  role?: Role;
+  active?: boolean;
+}
+
+export const useAccountStore = defineStore(
+  "account",
+  () => {
+    const account = ref<Account>({});
+
+    const isLoggedIn = computed(() =>
+      Boolean(account.value !== null && account.value.token)
+    );
+
+    // const isAdmin = computed(
+    //   () =>
+    //     account.value.role == Role.SuperAdmin ||
+    //     account.value.role == Role.Admin
+    // );
+    const isAdmin = true;
+
+    const mergeProfile = (profile: Partial<Account>) => {
+      if (account.value) {
+        Object.assign(account.value, profile);
+      }
+    };
+
+    const logout = () => {
+      account.value = {};
+    };
+
+    return { account, isLoggedIn, isAdmin, mergeProfile, logout };
   },
   {
     persist: true,