+++ tailwind.config.js
@@ -1,6 +1,7 @@
 /** @type {import('tailwindcss').Config} */
 export default {
   content: ["./src/**/*.{html,js,ts,vue}"],
+  darkMode: 'selector',
   theme: {
     extend: {},
   },